# backend-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Manajemen Anime Watchlist API",
        "description": "Dokumentasi lengkap REST API untuk Proyek Akhir Teknologi Integrasi Sistem. API ini menyediakan fitur autentikasi JWT, manajemen watchlist anime, pencarian anime via AniDB, dan manajemen user oleh admin.",
        "contact": {
            "name": "Raka Paksi",
            "email": "rakapaksi@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001/api",
            "description": "Backend Service Server"
        }
    ],
    "paths": {
        "/anime/search": {
            "get": {
                "tags": [
                    "Anime Search"
                ],
                "summary": "Search Anime",
                "description": "Mencari data anime dari external AniDB API berdasarkan keyword judul. Membutuhkan autentikasi JWT.",
                "operationId": "searchAnime",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Keyword pencarian judul anime",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Naruto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data anime berhasil diambil dari AniDB",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data anime dari AniDB berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "20"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Naruto"
                                                    },
                                                    "rating": {
                                                        "type": "string",
                                                        "example": "8.25"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "https://cdn.myanimelist.net/images/anime/13/17405.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Query pencarian kosong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Query pencarian tidak boleh kosong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid atau expired"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT yang didapat dari endpoint /login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}